Part 1 of 2 for the automatic pipeline creation refactor
Adds Fluent extensions for all methods on CommandList

This check in simplifies the automatic pipeline creation process. Needs to be refactored into a fluent syntax to eliminate the verbose argument list.

Also added two samples, one showing a simple game loop and an in-progress copy of the TexturedCube sample from Veldrid-Samples